requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.2
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
